Moving the ambient library code to the pkg folder so it follows standard golang layout
